requirements.txt
# requirements.txt
torch==2.7.0            # PyTorch for deep learning
transformers==4.51.3       # Hugging Face Transformers library
fastapi==0.115.12             # Web framework
uvicorn==0.34.2    # ASGI server
gunicorn==23.0.0          # WSGI server for FastAPI
sentence-transformers==4.1.0 # Hugging Face model library
pydantic==2.11.3            # Data validation (used by FastAPI)
pydantic-settings==2.9.1     # For loading settings from .env
python-dotenv==1.1.0       # Loads .env file for local development
sacremoses==0.1.1
sentencepiece==0.2.0
protobuf==6.30.2
psutil==5.9.8             # For system and process utilities (used in health check)
python-dotenv==1.1.0       # For loading environment variables from .env file